[improvement](assets): Add README and LICENSE
A README file contains information about other files in a directory or
archive of computer software. A form of documentation, it is usually a
simple plain text file called READ.ME, README.TXT, README.md,
README.1ST or simply README. (Source: Wikipedia)
Public software repositories are often used to share open source
software. For a repository to truly be open source, it's need to
license it so that others are free to use, change, and distribute the
software. (Source: help.github.com)