[maintain](resources): Add the project configuration
In the first step to port, it's necessary to properly configure your Makefile and workflow.xml, it's done by running "flyme config".
The tool adds the skeleton of the necessary files, and after that, the developer needs to edit them in order to fit his needs.
This commit does these two thing.